Automatically determine default container name when not given
# Background

Older versions of kubectl/oc, when ran as `kubectl logs -n <namespace>
<name>` first try to GET the pod resource requested, if they see there's
only one container in it, they will then try to GET:

`http://<api>/api/v1/namespaces/<namespace>/pods/<name>/log`

Without any `container=...` query param (newer versions of kubectl/oc
always specify this query parameter regardless). The behavior of the
regular kubernetes API server in this scenario is that when the pod has
only a single container, the logs for that only container are returned.
If the pod has more than one pod, an error response is returned, listing
the names of the containers in the pod.

# Issue

static-kas works fine when the container name is given in the query
param (which is the case for newer oc/kubectl clients), but breaks down
when the container parameter is omitted.

# Solution

static-kas will now better imitate the behavior of the regular
kubernetes API server